POPUG-023 [employee-billing] Refactoring: cleaned kafka-related packages.
