+++ employee-billing/src/main/java/org/uber/popug/employee/billing/config/KafkaConsumerConfig.java
@@ -13,16 +13,16 @@ import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
 import org.springframework.kafka.core.ConsumerFactory;
 import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
 import org.springframework.kafka.support.serializer.JsonDeserializer;
-import org.uber.popug.employee.billing.kafka.JsonSchemaRegistry;
-import org.uber.popug.employee.billing.kafka.TasksBusinessWorkflowJsonSchemaDeserializer;
-import org.uber.popug.employee.billing.kafka.TasksCUDJsonSchemaDeserializer;
-import org.uber.popug.employee.billing.kafka.consumer.JsonSchemaWithClass;
-import org.uber.popug.employee.billing.kafka.consumer.NamedJsonSchema;
+import org.uber.popug.employee.billing.kafka.common.JsonSchemaRegistry;
+import org.uber.popug.employee.billing.kafka.consumer.deserializer.TasksBusinessWorkflowJsonSchemaDeserializer;
+import org.uber.popug.employee.billing.kafka.consumer.deserializer.TasksCUDJsonSchemaDeserializer;
+import org.uber.popug.employee.billing.kafka.common.JsonSchemaWithClass;
+import org.uber.popug.employee.billing.kafka.common.NamedJsonSchema;
 import org.uber.popug.employee.billing.kafka.generated.dto.TaskCompletedEventV1;
 import org.uber.popug.employee.billing.kafka.generated.dto.TaskCreatedEventV1;
 import org.uber.popug.employee.billing.kafka.generated.dto.TaskCreatedReplicationEventV1;
 import org.uber.popug.employee.billing.kafka.generated.dto.TaskReassignedEventV1;
-import org.uber.popug.employee.billing.kafka.impl.CommonJsonSchemaRegistry;
+import org.uber.popug.employee.billing.kafka.common.impl.CommonJsonSchemaRegistry;
 
 import java.util.List;
 import java.util.Map;
