+++ employee-billing/src/main/java/org/uber/popug/employee/billing/kafka/common/NamedJsonSchema.java
@@ -1,4 +1,4 @@
-package org.uber.popug.employee.billing.kafka.consumer;
+package org.uber.popug.employee.billing.kafka.common;
 
 import jakarta.annotation.Nonnull;
