+++ employee-billing/src/main/java/org/uber/popug/employee/billing/kafka/common/impl/CommonJsonSchemaRegistry.java
@@ -1,7 +1,7 @@
-package org.uber.popug.employee.billing.kafka.impl;
+package org.uber.popug.employee.billing.kafka.common.impl;
 
-import org.uber.popug.employee.billing.kafka.JsonSchemaRegistry;
-import org.uber.popug.employee.billing.kafka.consumer.NamedJsonSchema;
+import org.uber.popug.employee.billing.kafka.common.JsonSchemaRegistry;
+import org.uber.popug.employee.billing.kafka.common.NamedJsonSchema;
 
 import java.util.Collection;
 import java.util.Map;
