+++ employee-billing/src/main/java/org/uber/popug/employee/billing/kafka/consumer/deserializer/TasksCUDJsonSchemaDeserializer.java
@@ -1,4 +1,6 @@
-package org.uber.popug.employee.billing.kafka;
+package org.uber.popug.employee.billing.kafka.consumer.deserializer;
+
+import org.uber.popug.employee.billing.kafka.common.JsonSchemaRegistry;
 
 public class TasksCUDJsonSchemaDeserializer extends HeaderBasedJsonSchemaDeserializer<Object> {
